welcome-clerk-7 Feature request - bootstrap security code scanning
detekt draft is found useful
klint not so much (use IDE)
