+++ .run/ApplicationKt.run.xml
@@ -1,7 +1,7 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="ApplicationKt" type="JetRunConfigurationType" nameIsGenerated="true">
-    <option name="MAIN_CLASS_NAME" value="th15.rdd13.me.ApplicationKt" />
-    <module name="th15.rdd13.me.welcome-clerk.main" />
+    <option name="MAIN_CLASS_NAME" value="me.rdd13.th15.ApplicationKt" />
+    <module name="me.rdd13.th15.welcome-clerk.main" />
     <shortenClasspath name="NONE" />
     <method v="2">
       <option name="Make" enabled="true" />
