+++ config/detekt.rdd13r.yml
@@ -1,0 +1,4 @@
+# Currently, no overrides needed.
+# See seeds at:
+# - https://github.com/detekt/detekt/blob/main/config/detekt/detekt.yml
+# - https://github.com/detekt/detekt/blob/main/config/detekt/statistics.yml
