+++ config/license.template
@@ -1,0 +1,5 @@
+/*
+MIT License
+Copyright (c) 2022 R!dd13r (rdd13r) - Vadim Kuhay
+See LICENSE
+ */
