--- src/main/kotlin/th15/rdd13/me/plugins/Routing.kt
@@ -1,19 +1,0 @@
-package th15.rdd13.me.plugins
-
-import io.ktor.server.application.*
-import io.ktor.server.response.*
-import io.ktor.server.routing.*
-
-fun Application.configureRouting() {
-
-    // Starting point for a Ktor app:
-    routing {
-        trace { application.log.trace(it.buildText()) }
-        get("/") {
-            call.respondText("Hello, I'm a 'Welcome' clerk, and my name is Välkommen!")
-        }
-    }
-
-    routing {
-    }
-}
